requirements.txt
kafka-python
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
fastapi
uvicorn
python-dotenv
requests
joblib
shap
sse-starlette
psycopg2-binary
SQLAlchemy
pyyaml
